Add geom_col, which makes bar plots using the identity stat
