+++ hail/python/hail/ggplot/__init__.py
@@ -1,8 +1,8 @@
 from .coord_cartesian import coord_cartesian
 from .ggplot import ggplot, GGPlot # noqa F401
 from .aes import aes, Aesthetic # noqa F401
-from .geoms import FigureAttribute, geom_line, geom_point, geom_text, geom_bar, geom_histogram, geom_hline, geom_func, geom_vline, \
-    geom_tile # noqa F401
+from .geoms import FigureAttribute, geom_line, geom_point, geom_text, geom_bar, geom_histogram, geom_func, \
+    geom_hline, geom_vline, geom_tile, geom_col # noqa F401
 from .labels import ggtitle, xlab, ylab
 from .scale import scale_x_continuous, scale_y_continuous, scale_x_discrete, scale_y_discrete, scale_x_genomic, \
     scale_x_log10, scale_y_log10, scale_x_reverse, scale_y_reverse, scale_color_discrete, scale_color_identity,\
@@ -15,6 +15,7 @@ __all__ = [
     "geom_line",
     "geom_text",
     "geom_bar",
+    "geom_col",
     "geom_histogram",
     "geom_hline",
     "geom_func",
